gh-65 Move all data flow resources
