+++ src/data-flows/index.ts
@@ -1,0 +1,3 @@
+export * from './mdm-data-class-component.resource';
+export * from './mdm-data-element-component.resource';
+export * from './mdm-data-flow.resource';
