+++ src/index.ts
@@ -21,6 +21,7 @@ export * from './mdm-common.resource';
 export * from './mdm-model-types.model';
 export * from './mdm-model-types.resource';
 export * from './mdm-container-types.model';
+export * from './data-flows/index';
 export * from './data-models/index';
 export * from './terminologies/index';
 export * from './reference-data/index';
@@ -43,14 +44,12 @@ export * from './mdm-admin.resource';
 export * from './mdm-rest-handler';
 export * from './mdm-security.model';
 export * from './mdm-security.resource';
-export * from './mdm-data-class-component.resource';
 export * from './mdm-namespaces.resource';
 export * from './mdm-catalogue-user.resource';
 export * from './mdm-session.model';
 export * from './mdm-session.resource';
 export * from './mdm-import-export.model';
 export * from './mdm-importer.resource';
-export * from './mdm-data-flow.resource';
 export * from './mdm-folder.model';
 export * from './mdm-folder.resource';
 export * from './mdm-tree-item.model';
@@ -58,7 +57,6 @@ export * from './mdm-tree-item.resource';
 export * from './mdm-metadata.resource';
 export * from './mdm-authentication.resource';
 export * from './mdm-user-image-file.resource';
-export * from './mdm-data-element-component.resource';
 export * from './mdm-edit.resource';
 export * from './mdm-resource';
 export * from './mdm-merge.model';
